check if exists before remove

+++ Depo/Depo/Business/HardwareService/MediaService/LocalMediaStorage.swift
@@ -774,11 +774,12 @@ class LocalMediaStorage: NSObject, LocalMediaStorageProtocol {
         let options = PHAssetResourceRequestOptions()
         options.isNetworkAccessAllowed = false
         
-        
-        
         let tmpUrl = Device.tmpFolderUrl(withComponent: name)
         
-        try? FileManager.default.removeItem(at: tmpUrl)
+        if FileManager.default.fileExists(atPath: tmpUrl.relativePath) {
+            debugLog("resumable_upload: cleaning tmp at \(tmpUrl.relativePath)")
+            try? FileManager.default.removeItem(at: tmpUrl)
+        }
         
         PHAssetResourceManager.default().writeData(for: resource, toFile: tmpUrl, options: options) { error in
             guard let error = error else {
